Kinda restarting
So for some reason, the last program written works aside from column
one, which randomly sets tick, speed, index and length to a value of 32
at random points in time, which seems to eventually lead to the
segfault. Fuck that noise.

Thought of a better way to do it anyway, so restarting with thematrix.c,
going to probably pull some code from main.c. Maybe not. Who knows.